Rename hook file to follow conventions
