+++ ui/src/hooks/vehicle-service/index.ts
@@ -1,1 +1,1 @@
-export * from './timetables';
+export * from './useGetTimetables';
